requirements.txt
duckdb
pandas
matplotlib.pyplot
seaborn
dbt-duckdb
